+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/utility/CameraTestingAutonomous.java
@@ -1,19 +1,19 @@
-package org.firstinspires.ftc.teamcode.autonomous.RI3W;
+package org.firstinspires.ftc.teamcode.autonomous.utility;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.hardware.RI3W.RI3WHardware;
-import org.firstinspires.ftc.teamcode.hardware.RI3W.vision.RI3WComputerVisionProcessor;
+import org.firstinspires.ftc.teamcode.hardware.vision.ComputerVisionProcessor;
 
 
 @Autonomous(name="Camera Test Auto", group="11588")
 public class CameraTestingAutonomous extends LinearOpMode {
     RI3WHardware robot = new RI3WHardware();
-    RI3WComputerVisionProcessor.PropPosition propPosition;
+    ComputerVisionProcessor.PropPosition propPosition;
     @Override
     public void runOpMode() throws InterruptedException {
-        robot.init(hardwareMap, telemetry, RI3WComputerVisionProcessor.AllianceColor.RED);
+        robot.init(hardwareMap, telemetry, ComputerVisionProcessor.AllianceColor.RED);
         waitForStart();
         while (!isStopRequested()) {
             autoLoop();
