+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/RI3W/RI3WHardware.java
@@ -2,9 +2,7 @@ package org.firstinspires.ftc.teamcode.hardware.RI3W;
 
 import com.qualcomm.hardware.bosch.BNO055IMU;
 import com.qualcomm.hardware.bosch.JustLoggingAccelerationIntegrator;
-import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.Servo;
 
@@ -14,9 +12,8 @@ import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
-import org.firstinspires.ftc.teamcode.hardware.RI3W.vision.RI3WComputerVisionProcessor;
+import org.firstinspires.ftc.teamcode.hardware.vision.ComputerVisionProcessor;
 import org.firstinspires.ftc.vision.VisionPortal;
-import org.firstinspires.ftc.vision.VisionProcessor;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraRotation;
 
@@ -52,19 +49,19 @@ public class RI3WHardware implements OpenCvCamera.AsyncCameraOpenListener {
     public DcMotorEx lift;
     public Servo claw;
     public BNO055IMU imu;
-    public RI3WComputerVisionProcessor pipeline;
+    public ComputerVisionProcessor pipeline;
 
     public OpenCvCamera camera;
     public VisionPortal visionPortal;
-    private RI3WComputerVisionProcessor visionProcessor;
+    private ComputerVisionProcessor visionProcessor;
 
     public void init(HardwareMap hardwareMap, Telemetry telemetry) {
-        init(hardwareMap, telemetry, RI3WComputerVisionProcessor.AllianceColor.BLUE);
+        init(hardwareMap, telemetry, ComputerVisionProcessor.AllianceColor.BLUE);
     }
-    public void init(HardwareMap hardwareMap, Telemetry telemetry, RI3WComputerVisionProcessor.AllianceColor allianceColor){
-        visionProcessor = new RI3WComputerVisionProcessor(telemetry, allianceColor);
+    public void init(HardwareMap hardwareMap, Telemetry telemetry, ComputerVisionProcessor.AllianceColor allianceColor){
+        visionProcessor = new ComputerVisionProcessor(telemetry, allianceColor);
 
-         pipeline = new RI3WComputerVisionProcessor(telemetry);
+         pipeline = new ComputerVisionProcessor(telemetry);
 
         // Create the vision portal the easy way.
         VisionPortal visionPortal = VisionPortal.easyCreateWithDefaults(
