+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/vision/CameraSubMat.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.hardware.RI3W.vision;
+package org.firstinspires.ftc.teamcode.hardware.vision;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.opencv.core.Core;
