+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleop/RI3W/RecordingTeleopRI3W.java
@@ -13,7 +13,7 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 
 @TeleOp(name="Recording Teleop", group="11588")
-public class RecordingTeleop extends OpMode {
+public class RecordingTeleopRI3W extends OpMode {
     public RI3WHardware robot = new RI3WHardware();
     public ElapsedTime timer = new ElapsedTime();
     public boolean isOpen = true;
@@ -30,7 +30,7 @@ public class RecordingTeleop extends OpMode {
 
     double turningPower = 0;
 
-    public RecordingTeleop() throws FileNotFoundException {
+    public RecordingTeleopRI3W() throws FileNotFoundException {
     }
 
     @Override
